% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fasttau}
\alias{fasttau}
\title{Kendall's tau correlation coeficient}
\usage{
fasttau(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
double
}
\description{
Computes Kendall's tau correlation coeficient through Rcpp.
}
\examples{
fasttau(c(1,2,3,4,5),c(2,1,4,3,6))
}
